* Added MirrorTaint jar with bug fixes * Added support for running benchmarks with Phosphor and MirrorTaint

+++ galette-benchmark/src/main/java/edu/neu/ccs/prl/galette/bench/extension/FlowReport.java
@@ -10,8 +10,11 @@ final class FlowReport {
     FlowReport(File report) throws IOException {
         Files.createDirectories(report.getParentFile().toPath());
         this.report = report;
-        try (PrintWriter out = new PrintWriter(new FileOutputStream(report, false))) {
-            out.println("class,method,tp,fp,fn,status");
+        boolean append = Boolean.getBoolean("flow.report.append");
+        if (!report.isFile() || !append) {
+            try (PrintWriter out = new PrintWriter(new FileOutputStream(report, false))) {
+                out.println("class,method,tp,fp,fn,status");
+            }
         }
     }
